WIP Gradle Build
This does not yet compile because some projects are missing
dependencies. The reason is that our previous Maven build was using
transitive dependencies a lot and did not stricly require modules to
declare all used dependencies. I did it differently in the Gradle build
where I declare almost everything as "implementation" for now to avoid
transitive dependencies from "leaking", i.e. projects that use a project
A cannot use the dependencies that project A declares as
"implementation".