[FLINK-19323][network] Small optimization of RecordWriter#serializeRecord
Currently, when serializing a record, the serializer will first skip 4 bytes for length filed and serialize the
record. Then it gets the serialized record length and skips back to position 0 to write the length field. After that,
it skip again to the tail of the serialized data. This patch avoids the last two skips by writing length field to
position 0 directly.

This closes #13447