[FLINK-21030][runtime] Adds trigger to do a global failover in case of failure
We cannot assume that the termination future is finishing when triggering a
synchronous savepoint. It might be that the savepoint creation succeeds but
the graceful shutdown of the job fails. In this case, the job termination does
not complete. The job might end up in an inconsistent state. That's why, a
global fail over is triggered if the scheduler observes one of the current
executions ending up in a non-finished state.

Four testcases were added testing the two different ways the
stop-with-savepoint can fail for the pre- and post-FLIP-27 source
implementations.